+++ src/channels/middleware/run-middleware.js
@@ -9,6 +9,7 @@ var mapWithPrefix = require('./middleware-utils').mapWithPrefix;
 module.exports = function (config, notifier) {
     var defaults = {
         serviceOptions: {},
+        initialOperation: [],
         variables: {
             autoFetch: true,
             silent: false,
@@ -23,7 +24,6 @@ module.exports = function (config, notifier) {
             autoFetch: true,
             readOnly: false
         },
-        initialOperation: '',
     };
     var opts = $.extend(true, {}, defaults, config);
 
@@ -38,16 +38,31 @@ module.exports = function (config, notifier) {
         $initialProm = $.Deferred().resolve(rs).promise();
     }
 
+    if (opts.initialOperation.length) {
+        //FIXME: Move run initialization logic to run-manager, as a strategy option. Technically only it should know what to do with it.
+        //For e.g, if there was a reset operation performed on the run, the run service instance will be the same so we wouldn't know
+        $initialProm = $initialProm.then(function (runService) {
+            if (!runService.initialize) {
+                runService.initialize = runService.serial(opts.initialOperation);
+            }
+            return runService.initialize.then(function () {
+                return runService;
+            });
+        });
+    }
+
     var variableschannel = new VariablesChannel();
+    //TODO: Need 2 different channel instances because the fetch is debounced, and hence will bundle variables up otherwise.
+    //also, notify needs to be called twice (with different arguments). Different way?
     var defaultVariablesChannel = new VariablesChannel();
     var metaChannel = new MetaChannel();
     var operationsChannel = new OperationsChannel();
 
     var handlers = [
-        $.extend(variableschannel, { name: 'variables', match: prefix('variable:') }),
-        $.extend(metaChannel, { name: 'meta', match: prefix('meta:') }),
-        $.extend(operationsChannel, { name: 'operations', match: prefix('operation:') }),
-        $.extend(defaultVariablesChannel, { name: 'variables', match: prefix('') }),
+        $.extend({}, variableschannel, { name: 'variables', match: prefix('variable:') }),
+        $.extend({}, metaChannel, { name: 'meta', match: prefix('meta:') }),
+        $.extend({}, operationsChannel, { name: 'operations', match: prefix('operation:') }),
+        $.extend({}, defaultVariablesChannel, { name: 'variables', match: prefix('') }),
     ];
 
     var notifyWithPrefix = function (prefix, data) {
@@ -83,7 +98,27 @@ module.exports = function (config, notifier) {
             });
         },
 
-        //TODO: Break this into multiple middlewares?
+        unsubscribeHandler: function (remainingTopics) {
+            handlers.reduce(function (pendingTopics, ph) {
+                var unsubs = ([].concat(pendingTopics)).reduce(function (accum, topic) {
+                    var prefixMatch = ph.match(topic, ph.prefix);
+                    if (prefixMatch !== false) {
+                        var stripped = topic.replace(prefixMatch, '');
+                        accum.myTopics.push(stripped);
+                        accum.prefix = prefixMatch;
+                    } else {
+                        accum.otherTopics.push(topic);
+                    }
+                    return accum;
+                }, { myTopics: [], otherTopics: [], prefix: '' });
+
+                if (unsubs.myTopics.length && ph.unsubscribeHandler) {
+                    ph.unsubscribeHandler(unsubs.myTopics);
+                }
+                return unsubs.otherTopics;
+            }, remainingTopics);
+        },
+
         publishHandler: function (inputObj) {
             return $initialProm.then(function (runService) {
                 //TODO: This means variables are always set before operations happen, make that more dynamic and by occurence order
@@ -118,8 +153,8 @@ module.exports = function (config, notifier) {
                         var unsilenced = silencable(resultObj, handlerOptions);
                         if (Object.keys(unsilenced).length && ph.name !== 'meta') {
                             //FIXME: Better way?
-                            variableschannel.fetch(runService).then(notifyWithPrefix.bind(null, 'variables:'));
-                            defaultVariablesChannel.fetch(runService).then(notifyWithPrefix.bind(null, ''));
+                            // variableschannel.fetch(runService).then(notifyWithPrefix.bind(null, 'variables:'));
+                            // defaultVariablesChannel.fetch(runService).then(notifyWithPrefix.bind(null, ''));
                         }
                         var mapped = mapWithPrefix(unsilenced, topicsToHandle.prefix);
                         return mapped;