+++ src/dom/attributes/attribute-manager.js
@@ -44,7 +44,7 @@
 
 var defaultHandlers = [
     require('./no-op-attr'),
-    require('./events/init-event-attr'),
+    // require('./events/init-event-attr'),
     require('./events/default-event-attr'),
     require('./foreach/default-foreach-attr'),
     require('./binds/checkbox-radio-bind-attr'),