+++ src/dom/attributes/repeat-attr.js
@@ -7,7 +7,7 @@
  *
  * 1. Add the `data-f-repeat` attribute to any HTML element that has repeated sub-elements. The two most common examples are lists and tables.
  * 2. Set the value of the `data-f-repeat` attribute in the HTML element you want to repeat to the name of the array variable.
- * 3. Optionally, inside the inner HTML element, use templates (`<%= %>`) to reference the `index` (for arrays) or `key` (for objects) and `value` to display. The `index`, `key`, and `value` are special variables that Flow.js populates for you.
+ * 3. Optionally, you can use templates (`<%= %>`) to reference the `index` (for arrays) or `key` (for objects) and `value` to display. The `index`, `key`, and `value` are special variables that Flow.js populates for you.
  *
  *
  * **Examples:**
@@ -42,43 +42,62 @@
  *          </tr>
  *      </table>
  *
+ * You can also use this with a `<div>` and have the `<div>` itself repeated. For example:
+ *
+ *      <div data-f-repeat="sample_array"></div>
+ *
+ * generates:
+ *
+ *      <div data-f-repeat="sample_array">2</div>
+ *      <div>4</div>
+ *      <div>6</div>
  *
  * **Notes:**
  *
- * * In most cases the same effect can be achieved with the [`data-f-foreach` attribute](../../attributes/foreach/default-foreach-attr/), which is similar. However, in the common use case of a table of data displayed over time, the `data-f-repeat` can be more concise and easier to read.
  * * You can use the `data-f-repeat` attribute with both arrays and objects. If the model variable is an object, reference the `key` instead of the `index` in your templates.
  * * The `key`, `index`, and `value` are special variables that Flow.js populates for you.
  * * The template syntax is to enclose each keyword (`index`, `key`, `variable`) in `<%=` and `%>`. Templates are available as part of Flow.js's lodash dependency. See more background on [working with templates](../../../../#templates).
+ * * In most cases the same effect can be achieved with the [`data-f-foreach` attribute](../../attributes/foreach/default-foreach-attr/), which is similar. In the common use case of a table of data displayed over time, the `data-f-repeat` can be more concise and easier to read. However, the `data-f-foreach` allows aliasing, and so can be more useful especially if you are nesting HTML elements or want to introduce logic about how to display the values.
  *
  */
 
 'use strict';
 var parseUtils = require('../../utils/parse-utils');
 var gutils = require('../../utils/general');
-// var config = require('../../config');
+var config = require('../../config').attrs;
 module.exports = {
 
     test: 'repeat',
 
     target: '*',
 
+    unbind: function (attr) {
+        var id = this.data(config.repeat.templateId);
+        if (id) {
+            this.nextUntil(':not([data-' + id + '])').remove();
+            // this.removeAttr('data-' + config.repeat.templateId);
+        }
+        var loopTemplate = this.data(config.repeat.template);
+        if (loopTemplate) {
+            this.removeData(config.repeat.template);
+            this.replaceWith(loopTemplate);
+        }
+    },
+
     handle: function (value, prop) {
         value = ($.isPlainObject(value) ? value : [].concat(value));
-        //FIXME: should ideally pull from config, but problem is, once it's unbound don't know what to remove
-        //Possible fixes: Let this handle it's own unbind (which does nothing), or
-        //have unbind remove all generated elements as well
-        var loopTemplate = this.data('repeat-template');
-        var id = this.data('repeat-template-id');
+        var loopTemplate = this.data(config.repeat.template);
+        var id = this.data(config.repeat.templateId);
 
         if (id) {
-            this.nextUntil(':not([data-' + id + '])').remove(); //clean-up pre-saved html
+            this.nextUntil(':not([data-' + id + '])').remove();
         } else {
             id = gutils.random('repeat-');
-            this.data('repeat-template-id', id);
+            this.data(config.repeat.templateId, id);
         }
         if (!loopTemplate) {
             loopTemplate = this.get(0).outerHTML;
-            this.data('repeat-template', loopTemplate);
+            this.data(config.repeat.template, loopTemplate);
         }
 
         var last;