+++ src/flow.js
@@ -69,6 +69,7 @@ var domManager = require('./dom/dom-manager');
 var BaseView = require('./utils/base-view');
 
 var ChannelManager = require('channels/channel-manager');
+var parseUtils = require('utils/parse-utils');
 
 var Flow = {
     dom: domManager,
@@ -80,10 +81,17 @@ var Flow = {
 
         var defaults = {
             channel: {
-                run: {
-                    account: '',
-                    project: '',
-                    model: model,
+                runManager: {
+                    run: {
+                        account: '',
+                        project: '',
+                        model: model,
+                    }
+                },
+                options: {
+                    runManager: {
+                        defaults: {}
+                    }
                 }
             },
             dom: {
@@ -95,7 +103,25 @@ var Flow = {
         var options = $.extend(true, {}, defaults, config);
         var $root = $(options.dom.root);
 
-        options.channel.initFn = $root.data('f-on-init');
+        var initialFn = $root.data('f-on-init');
+        //TOOD: Should move this to DOM Manager and just prioritize on-inits
+        if (initialFn) {
+            var listOfOperations = _.invoke(initialFn.split('|'), 'trim');
+            listOfOperations = listOfOperations.map(function (value) {
+                var fnName = value.split('(')[0];
+                var params = value.substring(value.indexOf('(') + 1, value.indexOf(')'));
+                var args = ($.trim(params) !== '') ? params.split(',') : [];
+                args = args.map(function (a) {
+                    return parseUtils.toImplicitType(a.trim());
+                });
+                var toReturn = {};
+                toReturn[fnName] = args;
+                return toReturn;
+            });
+
+            //TODO: Make a channel configuration factory which gets the initial info
+            options.channel.options.runManager.defaults.initialOperation = listOfOperations;
+        }
    
         if (config && config.channel && (config.channel instanceof ChannelManager)) {
             this.channel = config.channel;
@@ -108,7 +134,7 @@ var Flow = {
         }, options.dom));
     }
 };
-Flow.SubsManager = require('channels/subscription-manager');
+Flow.ChannelManager = ChannelManager;
 //set by grunt
-// if (RELEASE_VERSION) Flow.version = RELEASE_VERSION; //eslint-disable-line no-undef
+if (RELEASE_VERSION) Flow.version = RELEASE_VERSION; //eslint-disable-line no-undef
 module.exports = Flow;